+++ src/Labels/Labels.cpp
@@ -35,7 +35,7 @@ bool StatusNode::init()
     bottomRight->setID("bottom-right");
     this->addChild(bottomRight);
 
-    int count = 11;
+    int count = 12;
 
     for (size_t i = 0; i < count; i++)
     {
@@ -237,6 +237,9 @@ void StatusNode::update(float dt)
 
     if (!clock)
         clock = Client::GetModule("status-clock");
+
+    if (!percentage)
+        percentage = Client::GetModule("status-percentage");
         
     if (!attPL)
         attPL = static_cast<AttemptBaseGameLayer*>(GJBaseGameLayer::get());
@@ -262,6 +265,7 @@ void StatusNode::update(float dt)
     sLabels[8]->setVisible(cpsM->enabled);
     sLabels[9]->setVisible(bestRun->enabled);
     sLabels[10]->setVisible(clock->enabled);
+    sLabels[11]->setVisible(percentage->enabled);
 
     if (PlayLayer::get())
     {
@@ -334,6 +338,9 @@ void StatusNode::update(float dt)
         sLabels[9]->setString("Best Run: Editor");
 
     sLabels[10]->setString(formatTime().c_str());
+
+    if (bestRunPlayLayer && bestRunPlayLayer->m_percentageLabel)
+        sLabels[11]->setString(bestRunPlayLayer->m_percentageLabel->getString());
 
     updateVis();
 }